Fix Sauce Connect usage.
Sauce Connect appears to update the sessionId value but not the webdriver.remote.sessionid value
when proxying requests from the remote selenium instance to the local proxy.  The end result is
that, without this patch, php-webdriver can start an initial sesison but guesses the wrong URL
for subsequent requests.

sessionId in the result seems a more canonical place to get the sesion ID than the
webdrive.remote.sessionid key embedded in the value, so this patch amends AbstractWebDriver to
pass that through, using webdriver.remote.sessionid as a fail-over, just in case.

The change to WebDriver.php is a simple matter of pointing to the right variable.